FEAT: add login e logout

+++ src/app/services/local-storage.service.ts
@@ -1,25 +1,52 @@
 import { Injectable } from '@angular/core';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class LocalStorageService {
+  constructor() {}
 
-  constructor() { }
+  setItem(key: string, value: any) {
+    try {
+      if (typeof window !== 'undefined' && window.localStorage) {
+        localStorage.setItem(key, JSON.stringify(value));
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  }
 
   getItem(key: string) {
-    return localStorage.getItem(key);
+    try {
+      if (typeof window !== 'undefined' && window.localStorage) {
+        return localStorage.getItem(key);
+      }
+      // Add a return statement here
+      return null;
+    } catch (error) {
+      console.log(error);
+      // Add a return statement here
+      return null;
+    }
   }
 
   removeItem(key: string) {
-    localStorage.removeItem(key);
-  } 
-
-  clear() {
-    localStorage.clear();
+    try {
+      if (typeof window !== 'undefined' && window.localStorage) {
+        return localStorage.removeItem(key);
+      }
+    } catch (error) {
+      console.log(error);
+    }
   }
 
-  setItem(key: string, value: any) {
-    localStorage.setItem(key, JSON.stringify(value));
+  clear() {
+    try {
+      if (typeof window !== 'undefined' && window.localStorage) {
+        localStorage.clear();
+      }
+    } catch (error) {
+      console.log(error);
+    }
   }
 }
